Rough final patch for Gigabyte X299X DESIGNARE 10G
Note that this patch has not yet been adapted to use a quirk flag (or to be configuration-based).

This should not be used/ran on motherboards other than the Gigabyte X299X DESIGNARE 10G. However, if it is inadvertantly ran on other motherboards, it is unlikely to cause any significant issues.

This includes some commented debug code that may need to be removed for the finalized version (or perhaps it could be kept if it was gated to an even more verbose log level?). I've already stripped out almost all of the debug code that used to exist, including all of the unsafe WDT-related debug code.

IHandle.h was directly derived/excerpted from the UDK, and as such, there may be licensing issues with it. I preserved the original license notice from the file it was based on (this may possibly have been unnecessary). Note that simply referencing the entire original file will break things, hence the need for a partial version containing only the IHandle struct definition.